refactor(backtest): split domain.backtest into core/engine/io/ai/analysis
Behavior-preserving package reorg for readability (the flat 15-file package was
hard to navigate). Main source moved into subpackages by responsibility; tests
left in place with imports updated. No logic, signature, or visibility changes.
Full backtest+strategy suite green, identical test set.
