+++ src/test/kotlin/ru/driics/aitrade/domain/backtest/BacktestEngineTest.kt
@@ -3,6 +3,13 @@ package ru.driics.aitrade.domain.backtest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import ru.driics.aitrade.domain.backtest.core.Bar
+import ru.driics.aitrade.domain.backtest.core.ExitReason
+import ru.driics.aitrade.domain.backtest.core.PositionSide
+import ru.driics.aitrade.domain.backtest.engine.BacktestConfig
+import ru.driics.aitrade.domain.backtest.engine.BacktestEngine
+import ru.driics.aitrade.domain.backtest.engine.BacktestRejections
+import ru.driics.aitrade.domain.backtest.engine.InstrumentSpec
 import ru.driics.aitrade.domain.model.AiSignal
 import ru.driics.aitrade.domain.model.MarketState
 import ru.driics.aitrade.domain.strategy.Strategy
