+++ src/test/kotlin/ru/driics/aitrade/domain/backtest/BacktestRunnerReportTest.kt
@@ -3,6 +3,16 @@ package ru.driics.aitrade.domain.backtest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import ru.driics.aitrade.domain.backtest.core.ExitReason
+import ru.driics.aitrade.domain.backtest.core.PerformanceMetrics
+import ru.driics.aitrade.domain.backtest.core.PositionSide
+import ru.driics.aitrade.domain.backtest.core.SimTrade
+import ru.driics.aitrade.domain.backtest.engine.BacktestConfig
+import ru.driics.aitrade.domain.backtest.engine.BacktestRejections
+import ru.driics.aitrade.domain.backtest.engine.BacktestResult
+import ru.driics.aitrade.domain.backtest.engine.InstrumentSpec
+import ru.driics.aitrade.domain.backtest.io.BacktestReport
+import ru.driics.aitrade.domain.backtest.io.BacktestRunner
 import ru.driics.aitrade.domain.model.AiSignal
 import ru.driics.aitrade.domain.model.MarketState
 import ru.driics.aitrade.domain.strategy.Strategy
