+++ src/test/kotlin/ru/driics/aitrade/domain/backtest/JsonlCandleParserTest.kt
@@ -3,6 +3,7 @@ package ru.driics.aitrade.domain.backtest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertThrows
 import org.junit.jupiter.api.Test
+import ru.driics.aitrade.domain.backtest.io.JsonlCandleParser
 import java.math.BigDecimal
 
 class JsonlCandleParserTest {
